+++ ta4j-core/src/test/java/org/ta4j/core/indicators/helpers/DifferencePercentageIndicatorTest.java
@@ -25,7 +25,11 @@ package org.ta4j.core.indicators.helpers;
 
 import static org.ta4j.core.TestUtils.assertNumEquals;
 
+import java.time.Duration;
+import java.time.Instant;
+
 import org.junit.Test;
+import org.ta4j.core.BaseBarSeriesBuilder;
 import org.ta4j.core.Indicator;
 import org.ta4j.core.indicators.AbstractIndicatorTest;
 import org.ta4j.core.mocks.MockBarSeriesBuilder;
@@ -89,5 +93,141 @@ public class DifferencePercentageIndicatorTest extends AbstractIndicatorTest<Ind
         assertNumEquals(numOf(1), percentageChangeIndicator.getValue(2));
         assertNumEquals(numOf(2.5), percentageChangeIndicator.getValue(3));
         assertNumEquals(numOf(1.3), percentageChangeIndicator.getValue(4));
+    }
+
+    @Test
+    public void getValueWithOtherPreviousIndicator() {
+
+        var endTime = Instant.now();
+        var duration = Duration.ofSeconds(1);
+        var series = new BaseBarSeriesBuilder().withNumFactory(numFactory).build();
+
+        series.barBuilder()
+                .timePeriod(duration)
+                .endTime(endTime)
+                .openPrice(2.8)
+                .highPrice(1)
+                .lowPrice(1)
+                .closePrice(2.9)
+                .add();
+        series.barBuilder()
+                .timePeriod(duration)
+                .endTime(endTime.plusSeconds(1))
+                .openPrice(2.5)
+                .highPrice(2)
+                .lowPrice(2)
+                .closePrice(2.4)
+                .add();
+        series.barBuilder()
+                .timePeriod(duration)
+                .endTime(endTime.plusSeconds(2))
+                .openPrice(2.0)
+                .highPrice(1)
+                .lowPrice(1)
+                .closePrice(3.0)
+                .add();
+        series.barBuilder()
+                .timePeriod(duration)
+                .endTime(endTime.plusSeconds(3))
+                .openPrice(3)
+                .highPrice(3)
+                .lowPrice(3)
+                .closePrice(3.2)
+                .add();
+        series.barBuilder()
+                .timePeriod(duration)
+                .endTime(endTime.plusSeconds(4))
+                .openPrice(4)
+                .highPrice(4)
+                .lowPrice(4)
+                .closePrice(3.5)
+                .add();
+
+        var openPrice = new OpenPriceIndicator(series);
+        var closePrice = new ClosePriceIndicator(series);
+        var diff = new DifferencePercentageIndicator(openPrice, closePrice);
+
+        // index: 0: beginIndex <= index
+        assertNumEquals(NaN.NaN, diff.getValue(0));
+
+        // index: 1: currentOpenPrice = 2.5, previousClosePrice = 2.9
+        assertNumEquals(-13.79310344827586, diff.getValue(1));
+
+        // index: 2: currentOpenPrice = 2.0, previousClosePrice = 2.4
+        assertNumEquals(-16.666666666666666666666666666667, diff.getValue(2));
+
+        // index: 3: currentOpenPrice = 3.0, previousClosePrice = 3.0
+        assertNumEquals(0, diff.getValue(3));
+
+        // index: 4: currentOpenPrice = 4.0, previousClosePrice = 3.2
+        assertNumEquals(24.999999999999993, diff.getValue(4));
+    }
+
+    @Test
+    public void getValueWithOtherPreviousIndicatorAndThreshold() {
+
+        var endTime = Instant.now();
+        var duration = Duration.ofSeconds(1);
+        var series = new BaseBarSeriesBuilder().withNumFactory(numFactory).build();
+
+        series.barBuilder()
+                .timePeriod(duration)
+                .endTime(endTime)
+                .openPrice(2.8)
+                .highPrice(1)
+                .lowPrice(1)
+                .closePrice(2.9)
+                .add();
+        series.barBuilder()
+                .timePeriod(duration)
+                .endTime(endTime.plusSeconds(1))
+                .openPrice(2.5)
+                .highPrice(2)
+                .lowPrice(2)
+                .closePrice(2.4)
+                .add();
+        series.barBuilder()
+                .timePeriod(duration)
+                .endTime(endTime.plusSeconds(2))
+                .openPrice(2.0)
+                .highPrice(1)
+                .lowPrice(1)
+                .closePrice(3.0)
+                .add();
+        series.barBuilder()
+                .timePeriod(duration)
+                .endTime(endTime.plusSeconds(3))
+                .openPrice(3)
+                .highPrice(3)
+                .lowPrice(3)
+                .closePrice(3.2)
+                .add();
+        series.barBuilder()
+                .timePeriod(duration)
+                .endTime(endTime.plusSeconds(4))
+                .openPrice(4)
+                .highPrice(4)
+                .lowPrice(4)
+                .closePrice(3.5)
+                .add();
+
+        var openPrice = new OpenPriceIndicator(series);
+        var closePrice = new ClosePriceIndicator(series);
+        var diff = new DifferencePercentageIndicator(openPrice, closePrice, numOf(100));
+
+        // index: 0: beginIndex <= index
+        assertNumEquals(NaN.NaN, diff.getValue(0));
+
+        // index: 1: currentOpenPrice = 2.5, previousClosePrice = 2.9
+        assertNumEquals(-13.79310344827586, diff.getValue(1));
+
+        // index: 2: currentOpenPrice = 2.0, previousClosePrice = 2.4
+        assertNumEquals(-31.034482758620683, diff.getValue(2));
+
+        // index: 3: currentOpenPrice = 3.0, previousClosePrice = 3.0
+        assertNumEquals(3.448275862068968, diff.getValue(3));
+
+        // index: 4: currentOpenPrice = 4.0, previousClosePrice = 3.2
+        assertNumEquals(37.93103448275863, diff.getValue(4));
     }
 }